+++ nbed/scf/huzinaga_ks.py
@@ -5,12 +5,13 @@ from typing import Optional, Tuple
 
 import numpy as np
 import scipy as sp
+from pyscf.dft import rks, uks
 from pyscf.lib import StreamObject, diis
 
 logger = logging.getLogger(__name__)
 
 
-def huzinaga_RKS(
+def huzinaga_KS(
     scf_method: StreamObject,
     dft_potential: np.ndarray,
     dm_enviroment: np.ndarray,
@@ -18,7 +19,7 @@ def huzinaga_RKS(
     dm_initial_guess: Optional[np.ndarray] = None,
     use_DIIS: Optional[np.ndarray] = True,
 ) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, bool]:
-    """Manual RHF calculation that is implemented using the huzinaga operator.
+    """Manual RHF/UHF calculation that is implemented using the huzinaga operator.
 
     Note this function uses lowdin (symmetric) orthogonalization only! (PySCF sometimes uses meta-lowdin and NAO). Also
     the intial density matrix guess is based on the modified core Hamilotnian (containing projector and DFT potential)
@@ -27,7 +28,7 @@ def huzinaga_RKS(
     can occur due to DIIS and other clever PySCF methods not being available.
 
     Args:
-        scf_method (StreamObjecty):PySCF RHF object (containing info about max cycles and convergence tolerence)
+        scf_method (StreamObjecty):PySCF RHF/UHF object (containing info about max cycles and convergence tolerence)
         dft_potential (np.ndarray): DFT active and environment two body terms - DFT active environemnt two body term
         dm_enviroment (np.ndarray): Density matrix of the environment.
         dm_conv_tol (float): density matrix convergence tolerance
@@ -43,13 +44,33 @@ def huzinaga_RKS(
     s_mat = scf_method.get_ovlp()
     s_neg_half = sp.linalg.fractional_matrix_power(s_mat, -0.5)
 
+    if type(scf_method) is uks.UKS:
+        unrestricted = True
+    elif type(scf_method) is rks.RKS:
+        unrestricted = False
+    else:
+        logger.error("SCF method is not RKS or UKS")
+        raise TypeError("SCF method is not RKS or UKS")
+
+    if unrestricted:
+        dm_env_S = np.array([dm_enviroment[0] @ s_mat, dm_enviroment[1] @ s_mat])
+    else:
+        dm_env_S = dm_enviroment @ s_mat
+
     dm_env_S = dm_enviroment @ s_mat
     # Create an initial dm if needed.
     if dm_initial_guess is None:
         fock = scf_method.get_hcore() + dft_potential
 
-        fds = fock @ dm_env_S
-        huzinaga_op_std = -0.5 * (fds + fds.T)
+        if unrestricted:
+            fds_alpha = fock[0] @ dm_env_S[0]
+            fds_beta = fock[1] @ dm_env_S[1]
+            huzinaga_op_std = np.array(
+                [-(fds_alpha + fds_alpha.T), -(fds_beta + fds_beta.T)]
+            )
+        else:
+            fds = fock @ dm_env_S
+            huzinaga_op_std = -0.5 * (fds + fds.T)
 
         fock += huzinaga_op_std
         # Create the orthogonal fock operator
@@ -68,14 +89,19 @@ def huzinaga_RKS(
         adiis = diis.DIIS()
 
     for i in range(scf_method.max_cycle):
-
         # build fock matrix
         vhf = scf_method.get_veff(dm=dm_mat)
         fock = scf_method.get_hcore() + vhf + dft_potential
 
-        # projector
-        fds = fock @ dm_env_S
-        huzinaga_op_std = -0.5 * (fds + fds.T)
+        if unrestricted:
+            fds_alpha = fock[0] @ dm_env_S[0]
+            fds_beta = fock[1] @ dm_env_S[1]
+            huzinaga_op_std = np.array(
+                [-(fds_alpha + fds_alpha.T), -(fds_beta + fds_beta.T)]
+            )
+        else:
+            fds = fock @ dm_env_S
+            huzinaga_op_std = -0.5 * (fds + fds.T)
 
         fock += huzinaga_op_std
 
@@ -95,15 +121,31 @@ def huzinaga_RKS(
         # Find RKS energy
         #     rks_energy = scf_method.energy_elec(dm=dm_mat)[0]
         vhf_updated = scf_method.get_veff(dm=dm_mat)
-        rks_energy = (
-            vhf_updated.ecoul
-            + vhf_updated.exc
-            + np.einsum(
-                "ij, ji->",
-                dm_mat,
-                (scf_method.get_hcore() + huzinaga_op_std + dft_potential),
+        if unrestricted:
+            rks_energy = (
+                vhf_updated.ecoul
+                + vhf_updated.exc
+                + np.einsum(
+                    "ij, ji->",
+                    dm_mat[0],
+                    (scf_method.get_hcore()[0] + huzinaga_op_std[0] + dft_potential[0]),
+                )
+                + np.einsum(
+                    "ij, ji->",
+                    dm_mat[1],
+                    (scf_method.get_hcore()[1] + huzinaga_op_std[1] + dft_potential[1]),
+                )
             )
-        )
+        else:
+            rks_energy = (
+                vhf_updated.ecoul
+                + vhf_updated.exc
+                + np.einsum(
+                    "ij, ji->",
+                    dm_mat,
+                    (scf_method.get_hcore() + huzinaga_op_std + dft_potential),
+                )
+            )
 
         # check convergence
         run_diff = np.abs(rks_energy - rks_energy_prev)